[GraphQL/Package] chore: Move Package Resolver implementation (#17178)
## Description

Various clean-ups related to GraphQL's package resolver implementation:

- Move from `context_data::package_cache` to `data::package_resolver`.
This is to acknowledge the fact that it's no longer going to be limited
to just the context data (because it also needs to be shared with the
task that evicts it on epoch boundaries).
- Inline the various fetching APIs.
- Introduce an alias for the resolver type itself, as the package cache
alias was only ever used as a type parameter to `Resolver`.

These are in preparation for changing its implementation to use a task
for evicting system packages on epoch boundaries, rather than the
current system that does a version check on each fetch that involves a
system package.

## Test plan

Existing tests as this is a simple refactoring.

## Stack

- #17177

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol:
- [ ] Nodes (Validators and Full nodes):
- [ ] Indexer:
- [ ] JSON-RPC:
- [ ] GraphQL:
- [ ] CLI:
- [ ] Rust SDK:
