+++ crates/sui-graphql-rpc/src/context_data/mod.rs
@@ -2,4 +2,3 @@
 // SPDX-License-Identifier: Apache-2.0
 
 pub(crate) mod db_data_provider;
-pub(crate) mod package_cache;
