+++ crates/sui-graphql-rpc/src/data.rs
@@ -1,6 +1,7 @@
 // Copyright (c) Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
+pub(crate) mod package_resolver;
 pub(crate) mod pg;
 
 use async_trait::async_trait;
